fix: #590 add Kernel.onActivate - which is called after gpu.js switches kernels based off need fix: Consistent error message for values that are not defined fix: Add recompiled kernels to gpu.js kernels property and test fix: Move texture deleting to within the GLKernel Texture implementation fix: Add Texture.clear as an abstract method on the base Texture fix: Bump and build

+++ src/backend/gl/texture/index.js
@@ -87,7 +87,13 @@ class GLTexture extends Texture {
   }
 
   delete() {
-    super.delete();
+    if (this._deleted) return;
+    this._deleted = true;
+    if (this.texture._refs) {
+      this.texture._refs--;
+      if (this.texture._refs) return;
+    }
+    this.context.deleteTexture(this.texture);
     if (this.texture._refs === 0 && this._framebuffer) {
       this.context.deleteFramebuffer(this._framebuffer);
       this._framebuffer = null;
